split up BloopExporter
