+++ commons/src/main/java/org/jetbrains/bsp/bazel/commons/Constants.java
@@ -23,7 +23,7 @@ public class Constants {
   public static final String BAZELBSP_JSON_FILE_NAME = "bazelbsp.json";
   public static final String SERVER_CLASS_NAME = "org.jetbrains.bsp.bazel.server.ServerInitializer";
   public static final String BLOOP_BOOTSTRAP_CLASS_NAME =
-      "org.jetbrains.bsp.bazel.server.BloopExporterInitializer";
+      "org.jetbrains.bsp.bazel.server.bloop.BloopExporterInitializer";
   public static final String BLOOP_SETTINGS_JSON_FILE_NAME = "bloop.settings.json";
   public static final String CLASSPATH_FLAG = "-classpath";
   public static final String BAZELBSP_TRACE_JSON_FILE_NAME = "bazelbsp.trace.json";
