+++ server/src/main/java/org/jetbrains/bsp/bazel/server/bloop/BUILD
@@ -1,7 +1,7 @@
 load("@rules_java//java:defs.bzl", "java_binary", "java_library")
 
 java_library(
-    name = "server",
+    name = "bloop",
     srcs = glob(["*.java"]),
     visibility = ["//server:__subpackages__"],
     deps = [
@@ -10,7 +10,6 @@ java_library(
         "//executioncontext/workspacecontext",
         "//logger",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bep",
-        "//server/src/main/java/org/jetbrains/bsp/bazel/server/bloop",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bsp",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/info",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/managers",
