+++ server/src/main/java/org/jetbrains/bsp/bazel/server/common/BUILD
@@ -1,7 +1,7 @@
 load("@rules_java//java:defs.bzl", "java_binary", "java_library")
 
 java_library(
-    name = "server",
+    name = "common",
     srcs = glob(["*.java"]),
     visibility = ["//server:__subpackages__"],
     deps = [
@@ -10,13 +10,11 @@ java_library(
         "//executioncontext/workspacecontext",
         "//logger",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bep",
-        "//server/src/main/java/org/jetbrains/bsp/bazel/server/bloop",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bsp",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/info",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/managers",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/services",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/utils",
-        "//server/src/main/java/org/jetbrains/bsp/bazel/server/common",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/diagnostics",
         "//server/src/main/java/org/jetbrains/bsp/bazel/server/sync",
         "@io_bazel//src/main/protobuf:build_java_proto",
