+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/languages/scala/ScalaModule.java
@@ -9,6 +9,14 @@ import org.jetbrains.bsp.bazel.server.sync.languages.LanguageData;
 import org.jetbrains.bsp.bazel.server.sync.languages.java.JavaModule;
 
 public class ScalaModule implements LanguageData {
+  public static Option<ScalaModule> fromLanguageData(LanguageData languageData) {
+    if (languageData instanceof ScalaModule) {
+      return Option.of((ScalaModule) languageData);
+    }
+
+    return Option.none();
+  }
+
   private final ScalaSdk sdk;
   private final Seq<String> scalacOpts;
   private final Option<JavaModule> javaModule;
